requirements.txt
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
imageio==2.37.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.6.0
typing_extensions==4.13.2
